requirements.txt
requests
pandas
google-cloud-firestore
google-auth
scikit-learn
